+++ src/form/bf_mx_sdp.jl
@@ -6,8 +6,8 @@ function constraint_mc_model_current(pm::SDPUBFModel, n::Int, i, f_bus, f_idx, g
     w_fr_re = _PMs.var(pm, n, :Wr)[f_bus]
     w_fr_im = _PMs.var(pm, n, :Wi)[f_bus]
 
-    ccm_re =  _PMs.var(pm, n, :CC_re)[i]
-    ccm_im =  _PMs.var(pm, n, :CC_im)[i]
+    ccm_re =  _PMs.var(pm, n, :CCr)[i]
+    ccm_im =  _PMs.var(pm, n, :CCi)[i]
 
     p_s_fr = p_fr - (w_fr_re*(g_sh_fr)' + w_fr_im*(b_sh_fr)')
     q_s_fr = q_fr - (w_fr_im*(g_sh_fr)' - w_fr_re*(b_sh_fr)')
