+++ src/form/bf_mx_soc.jl
@@ -1,13 +1,13 @@
 "Defines relationship between branch (series) power flow, branch (series) current and node voltage magnitude"
 function constraint_mc_model_current(pm::SOCUBFModels, n::Int, i, f_bus, f_idx, g_sh_fr, b_sh_fr)
-    p_fr = _PMs.var(pm, n, :P_mx)[f_idx]
-    q_fr = _PMs.var(pm, n, :Q_mx)[f_idx]
+    p_fr = _PMs.var(pm, n, :P)[f_idx]
+    q_fr = _PMs.var(pm, n, :Q)[f_idx]
 
     w_fr_re = _PMs.var(pm, n, :Wr)[f_bus]
     w_fr_im = _PMs.var(pm, n, :Wi)[f_bus]
 
-    ccm_re =  _PMs.var(pm, n, :CC_re)[i]
-    ccm_im =  _PMs.var(pm, n, :CC_im)[i]
+    ccm_re =  _PMs.var(pm, n, :CCr)[i]
+    ccm_im =  _PMs.var(pm, n, :CCi)[i]
 
     p_s_fr = p_fr - g_sh_fr*w_fr_re
     q_s_fr = q_fr + b_sh_fr*w_fr_re
@@ -36,14 +36,14 @@ end
 
 "Defines relationship between branch (series) power flow, branch (series) current and node voltage magnitude"
 function constraint_mc_model_current(pm::SOCConicUBFModel, n::Int, i, f_bus, f_idx, g_sh_fr, b_sh_fr)
-    p_fr = _PMs.var(pm, n, :P_mx)[f_idx]
-    q_fr = _PMs.var(pm, n, :Q_mx)[f_idx]
+    p_fr = _PMs.var(pm, n, :P)[f_idx]
+    q_fr = _PMs.var(pm, n, :Q)[f_idx]
 
     w_fr_re = _PMs.var(pm, n, :Wr)[f_bus]
     w_fr_im = _PMs.var(pm, n, :Wi)[f_bus]
 
-    ccm_re =  _PMs.var(pm, n, :CC_re)[i]
-    ccm_im =  _PMs.var(pm, n, :CC_im)[i]
+    ccm_re =  _PMs.var(pm, n, :CCr)[i]
+    ccm_im =  _PMs.var(pm, n, :CCi)[i]
 
     p_s_fr = p_fr - g_sh_fr*w_fr_re
     q_s_fr = q_fr + b_sh_fr*w_fr_re
