+++ src/prob/tp_opf_bf_del.jl
@@ -1,23 +1,23 @@
 ""
-function run_tp_opf_bf_del(data::Dict{String,Any}, model_constructor, solver; kwargs...)
-    return _PMs.run_model(data, model_constructor, solver, post_tp_opf_bf_del; solution_builder=solution_tp!, multiconductor=true, kwargs...)
+function run_mc_opf_bf_del(data::Dict{String,Any}, model_constructor, solver; kwargs...)
+    return _PMs.run_model(data, model_constructor, solver, post_mc_opf_bf_del; solution_builder=solution_tp!, multiconductor=true, kwargs...)
 end
 
 
 ""
-function run_tp_opf_bf_del(file::String, model_constructor, solver; kwargs...)
-    return run_tp_opf_bf_del(PowerModelsDistribution.parse_file(file), model_constructor, solver; kwargs...)
+function run_mc_opf_bf_del(file::String, model_constructor, solver; kwargs...)
+    return run_mc_opf_bf_del(PowerModelsDistribution.parse_file(file), model_constructor, solver; kwargs...)
 end
 
 
 ""
-function post_tp_opf_bf_del(pm::_PMs.GenericPowerModel)
+function post_mc_opf_bf_del(pm::_PMs.GenericPowerModel)
     # Variables
-    variable_tp_voltage(pm)
-    variable_tp_branch_current(pm)
-    variable_tp_branch_flow(pm)
+    variable_mc_voltage(pm)
+    variable_mc_branch_current(pm)
+    variable_mc_branch_flow(pm)
 
-    variable_tp_load(pm)
+    variable_mc_load(pm)
 
     for c in _PMs.conductor_ids(pm)
         _PMs.variable_generation(pm, cnd=c)
@@ -25,16 +25,16 @@ function post_tp_opf_bf_del(pm::_PMs.GenericPowerModel)
     end
 
     # Constraints
-    constraint_tp_model_current(pm)
+    constraint_mc_model_current(pm)
 
     for i in _PMs.ids(pm, :ref_buses)
-        constraint_tp_theta_ref(pm, i)
+        constraint_mc_theta_ref(pm, i)
     end
 
     for i in _PMs.ids(pm, :branch)
-        constraint_tp_flow_losses(pm, i)
+        constraint_mc_flow_losses(pm, i)
 
-        constraint_tp_model_voltage_magnitude_difference(pm, i)
+        constraint_mc_model_voltage_magnitude_difference(pm, i)
 
         for c in _PMs.conductor_ids(pm)
             #_PMs.constraint_voltage_angle_difference(pm, i, cnd=c)
@@ -45,7 +45,7 @@ function post_tp_opf_bf_del(pm::_PMs.GenericPowerModel)
     end
 
     for i in _PMs.ids(pm, :load)
-        constraint_tp_load(pm, i)
+        constraint_mc_load(pm, i)
     end
 
     for i in _PMs.ids(pm, :dcline), c in _PMs.conductor_ids(pm)
@@ -53,7 +53,7 @@ function post_tp_opf_bf_del(pm::_PMs.GenericPowerModel)
     end
 
     for i in _PMs.ids(pm, :bus), c in _PMs.conductor_ids(pm)
-        constraint_power_balance_shunt(pm, i, cnd=c)
+        constraint_mc_power_balance_shunt(pm, i, cnd=c)
     end
 
     # Objective
